--- src/pages/Order/components/ORDERED_TABLE_TITLE.js
@@ -1,7 +1,0 @@
-export const ORDERED_TABLE_TITLE = [
-  { id: 1, title: '주문번호' },
-  { id: 2, title: '상품명' },
-  { id: 3, title: '수량' },
-  { id: 4, title: '배송지' },
-  { id: 5, title: '수령희망일' },
-];
