+++ src/pages/Ordered/components/OrderedInfo.js
